some cleanup

+++ common/primitives/src/node.rs
@@ -37,7 +37,7 @@ pub type Signature = MultiSignature;
 pub type Index = u32;
 
 /// the type for the index of the Staking Reward time period
-pub type RewardEra = u64;
+pub type RewardEra = u32;
 
 /// A hash of some data used by the chain.
 pub type Hash = sp_core::H256;
